Search complete, home complete, ion card WIP
Search now is able to search the api with each enter of a new letter.
home displays the top 40 games of the api and has a next button that gets the next 40 games for a casual browse.
Ion card missing some features for expanding or at least opening the card for more details and better formatting